Implements SP-106 tasks for FHIRPath specification compliance:
- Quantity literal support (10 'mg') with FHIR Quantity JSON
- DateTime type checking for @YYYYT vs @YYYY (Date vs DateTime)
- Decimal precision fix (DOUBLE -> DECIMAL in DuckDB)
- Timezone validation for time literals (@T14:34:28Z rejected)
- subsetOf/supersetOf column preservation
- Unary polarity operators (already working)
- FHIR Quantity business logic moved from translator to types module
- Consolidated duplicate datetime type checking logic
- Added 94 new tests (75 passing)

Compliance: 57.4% → 58.6% (+11 tests)

+++ fhir4ds/main/dialects/duckdb.py
@@ -1405,10 +1405,11 @@ class DuckDBDialect(DatabaseDialect):
         """
         # Map FHIRPath types to DuckDB type names (uppercase)
         # This mapping is part of syntax adaptation, not business logic
+        # SP-106-004: Use DECIMAL for FHIRPath decimal type to maintain precision
         type_map = {
             "string": "VARCHAR",
             "integer": "INTEGER",
-            "decimal": "DOUBLE",
+            "decimal": "DECIMAL",
             "boolean": "BOOLEAN",
             "datetime": "TIMESTAMP",
             "date": "DATE",
